requirements.txt
# requirements.txt generated by rsconnect-python on 2025-02-20 09:59:15.778132+00:00
altair==4.1.0
pandas==1.4.4
streamlit==1.25.0
numpy
scipy
scikit-learn==1.3.2
ipython
matplotlib
scikit-image
kaleido
reportlab
pillow
joblib==1.3.2
xlsxwriter
PyPDF2
openpyxl
pycryptodome
python-docx